+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 - New view for Files/Datasets offering a Table View of the attached Metadata.
+- Add SUBMITTED event on the GUI of extractions and pass this submitted event id into the rabbitmq message.
+  [CATS-969] (https://opensource.ncsa.illinois.edu/jira/browse/CATS-969)
 - Pass user email address in the extraction rabbitmq message when file uploading and manually submitting extraction.
   [CATS-963](https://opensource.ncsa.illinois.edu/jira/browse/CATS-963)
 - Extraction history for dataset extractors is now displayed on dataset page.